+++ Origo.Core/DataSource/Converters/PrimitiveConverters.cs
@@ -11,7 +11,7 @@ internal sealed class StringDataSourceConverter : DataSourceConverter<string>
 
 internal sealed class ByteDataSourceConverter : DataSourceConverter<byte>
 {
-    public override byte Read(DataSourceNode node) => node.AsByte();
+    public override byte Read(DataSourceNode node) => node.As<byte>();
 
     public override DataSourceNode Write(byte value) =>
         DataSourceNode.CreateNumber(value.ToString(CultureInfo.InvariantCulture));
@@ -19,7 +19,7 @@ internal sealed class ByteDataSourceConverter : DataSourceConverter<byte>
 
 internal sealed class SByteDataSourceConverter : DataSourceConverter<sbyte>
 {
-    public override sbyte Read(DataSourceNode node) => node.AsSByte();
+    public override sbyte Read(DataSourceNode node) => node.As<sbyte>();
 
     public override DataSourceNode Write(sbyte value) =>
         DataSourceNode.CreateNumber(value.ToString(CultureInfo.InvariantCulture));
@@ -27,7 +27,7 @@ internal sealed class SByteDataSourceConverter : DataSourceConverter<sbyte>
 
 internal sealed class Int16DataSourceConverter : DataSourceConverter<short>
 {
-    public override short Read(DataSourceNode node) => node.AsShort();
+    public override short Read(DataSourceNode node) => node.As<short>();
 
     public override DataSourceNode Write(short value) =>
         DataSourceNode.CreateNumber(value.ToString(CultureInfo.InvariantCulture));
@@ -35,7 +35,7 @@ internal sealed class Int16DataSourceConverter : DataSourceConverter<short>
 
 internal sealed class UInt16DataSourceConverter : DataSourceConverter<ushort>
 {
-    public override ushort Read(DataSourceNode node) => node.AsUShort();
+    public override ushort Read(DataSourceNode node) => node.As<ushort>();
 
     public override DataSourceNode Write(ushort value) =>
         DataSourceNode.CreateNumber(value.ToString(CultureInfo.InvariantCulture));
@@ -43,14 +43,14 @@ internal sealed class UInt16DataSourceConverter : DataSourceConverter<ushort>
 
 internal sealed class Int32DataSourceConverter : DataSourceConverter<int>
 {
-    public override int Read(DataSourceNode node) => node.AsInt();
+    public override int Read(DataSourceNode node) => node.As<int>();
 
     public override DataSourceNode Write(int value) => DataSourceNode.CreateNumber(value);
 }
 
 internal sealed class UInt32DataSourceConverter : DataSourceConverter<uint>
 {
-    public override uint Read(DataSourceNode node) => node.AsUInt();
+    public override uint Read(DataSourceNode node) => node.As<uint>();
 
     public override DataSourceNode Write(uint value) =>
         DataSourceNode.CreateNumber(value.ToString(CultureInfo.InvariantCulture));
@@ -58,14 +58,14 @@ internal sealed class UInt32DataSourceConverter : DataSourceConverter<uint>
 
 internal sealed class Int64DataSourceConverter : DataSourceConverter<long>
 {
-    public override long Read(DataSourceNode node) => node.AsLong();
+    public override long Read(DataSourceNode node) => node.As<long>();
 
     public override DataSourceNode Write(long value) => DataSourceNode.CreateNumber(value);
 }
 
 internal sealed class UInt64DataSourceConverter : DataSourceConverter<ulong>
 {
-    public override ulong Read(DataSourceNode node) => node.AsULong();
+    public override ulong Read(DataSourceNode node) => node.As<ulong>();
 
     public override DataSourceNode Write(ulong value) =>
         DataSourceNode.CreateNumber(value.ToString(CultureInfo.InvariantCulture));
@@ -73,21 +73,21 @@ internal sealed class UInt64DataSourceConverter : DataSourceConverter<ulong>
 
 internal sealed class SingleDataSourceConverter : DataSourceConverter<float>
 {
-    public override float Read(DataSourceNode node) => node.AsFloat();
+    public override float Read(DataSourceNode node) => node.As<float>();
 
     public override DataSourceNode Write(float value) => DataSourceNode.CreateNumber(value);
 }
 
 internal sealed class DoubleDataSourceConverter : DataSourceConverter<double>
 {
-    public override double Read(DataSourceNode node) => node.AsDouble();
+    public override double Read(DataSourceNode node) => node.As<double>();
 
     public override DataSourceNode Write(double value) => DataSourceNode.CreateNumber(value);
 }
 
 internal sealed class DecimalDataSourceConverter : DataSourceConverter<decimal>
 {
-    public override decimal Read(DataSourceNode node) => node.AsDecimal();
+    public override decimal Read(DataSourceNode node) => node.As<decimal>();
 
     public override DataSourceNode Write(decimal value) =>
         DataSourceNode.CreateNumber(value.ToString(CultureInfo.InvariantCulture));
@@ -102,7 +102,7 @@ internal sealed class CharDataSourceConverter : DataSourceConverter<char>
 
 internal sealed class BooleanDataSourceConverter : DataSourceConverter<bool>
 {
-    public override bool Read(DataSourceNode node) => node.AsBool();
+    public override bool Read(DataSourceNode node) => node.As<bool>();
 
     public override DataSourceNode Write(bool value) => DataSourceNode.CreateBoolean(value);
 }